优化&更新依赖
