+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "lx-music-desktop",
-  "version": "2.3.0-beta.5",
+  "version": "2.3.0-beta.6",
   "description": "一个免费的音乐查找助手",
   "main": "./dist/main.js",
   "productName": "lx-music-desktop",
@@ -230,11 +230,11 @@
     "css-loader": "^6.8.1",
     "css-minimizer-webpack-plugin": "^5.0.0",
     "del": "^6.1.1",
-    "electron": "^22.3.11",
+    "electron": "^22.3.12",
     "electron-builder": "^24.4.0",
     "electron-debug": "^3.2.0",
     "electron-devtools-installer": "^3.2.0",
-    "electron-to-chromium": "^1.4.413",
+    "electron-to-chromium": "^1.4.416",
     "electron-updater": "^6.1.0",
     "eslint": "^8.41.0",
     "eslint-config-standard": "^17.1.0",
@@ -249,7 +249,7 @@
     "eslint-webpack-plugin": "^4.0.1",
     "html-webpack-plugin": "^5.5.1",
     "less": "^4.1.3",
-    "less-loader": "^11.1.1",
+    "less-loader": "^11.1.2",
     "mini-css-extract-plugin": "^2.7.6",
     "node-loader": "^2.0.0",
     "postcss": "^8.4.24",
@@ -262,14 +262,14 @@
     "svg-sprite-loader": "^6.0.11",
     "svg-transform-loader": "^2.0.13",
     "svgo-loader": "^4.0.0",
-    "terser": "^5.17.6",
+    "terser": "^5.17.7",
     "terser-webpack-plugin": "^5.3.9",
     "ts-loader": "^9.4.3",
     "typescript": "^5.0.4",
     "vue-eslint-parser": "^9.3.0",
-    "vue-loader": "^17.1.2",
+    "vue-loader": "^17.2.1",
     "vue-template-compiler": "^2.7.14",
-    "webpack": "^5.84.1",
+    "webpack": "^5.85.0",
     "webpack-cli": "^5.1.1",
     "webpack-dev-server": "^4.15.0",
     "webpack-hot-middleware": "github:lyswhut/webpack-hot-middleware#329c4375134b89d39da23a56a94db651247c74a1",
