+++ publish/changeLog.md
@@ -13,4 +13,4 @@
 
 ### 其他
 
-- 更新 electron 到 v22.3.11
+- 更新 electron 到 v22.3.12
