+++ src/renderer/plugins/player/index.ts
@@ -39,7 +39,7 @@ export const convolutions = [
   // { name: 'tim_omni_rear_blend', mainGain: 1.8, sendGain: 0.8, source: 'tim-omni-rear-blend.wav' },
 ] as const
 // 半音
-export const semitones = [-1.5, -1, -0.5, 0.5, 1, 1.5, 2, 2.5, 3, 3.5] as const
+// export const semitones = [-1.5, -1, -0.5, 0.5, 1, 1.5, 2, 2.5, 3, 3.5] as const
 
 let convolver: ConvolverNode
 let convolverSourceGainNode: GainNode
@@ -48,7 +48,7 @@ let convolverDynamicsCompressor: DynamicsCompressorNode
 let gainNode: GainNode
 let panner: PannerNode
 let pitchShifterNode: AudioWorkletNode
-let pitchShifterNodeAudioParam: AudioParam
+let pitchShifterNodePitchFactor: AudioParam
 let pitchShifterNodeLoadStatus: 'none' | 'loading' | 'unconnect' | 'connected' = 'none'
 let pitchShifterNodeTempValue = 1
 export const soundR = 0.5
@@ -225,14 +225,14 @@ const loadPitchShifterNode = () => {
   // source -> analyser -> biquadFilter -> audioWorklet(pitch shifter) -> [(convolver & convolverSource)->convolverDynamicsCompressor] -> panner -> gain
   void audioContext.audioWorklet.addModule(new URL(
     /* webpackChunkName: 'pitch_shifter.audioWorklet' */
-    '@renderer/utils/pitch-shifter/phase-vocoder.js',
+    './pitch-shifter/phase-vocoder.js',
     import.meta.url,
   )).then(() => {
     console.log('pitch shifter audio worklet loaded')
     pitchShifterNode = new AudioWorkletNode(audioContext, 'phase-vocoder-processor')
-    let audioParam = pitchShifterNode.parameters.get('pitchFactor')
-    if (!audioParam) return
-    pitchShifterNodeAudioParam = audioParam
+    let pitchFactorParam = pitchShifterNode.parameters.get('pitchFactor')
+    if (!pitchFactorParam) return
+    pitchShifterNodePitchFactor = pitchFactorParam
     pitchShifterNodeLoadStatus = 'unconnect'
     if (pitchShifterNodeTempValue == 1) return
 
@@ -250,7 +250,7 @@ const connectPitchShifterNode = () => {
   // convolverDynamicsCompressor.connect(pitchShifterNode)
   // pitchShifterNode.connect(panner)
   pitchShifterNodeLoadStatus = 'connected'
-  pitchShifterNodeAudioParam.value = pitchShifterNodeTempValue
+  pitchShifterNodePitchFactor.value = pitchShifterNodeTempValue
 }
 // const disconnectPitchShifterNode = () => {
 //   const lastBiquadFilter = (biquads.get(`hz${freqs.at(-1) as Freqs}`) as BiquadFilterNode)
@@ -275,7 +275,7 @@ export const setPitchShifter = (val: number) => {
     case 'connected':
       // a: 1 = 半音
       // value = 2 ** (a / 12)
-      pitchShifterNodeAudioParam.value = val
+      pitchShifterNodePitchFactor.value = val
       break
     case 'unconnect':
       connectPitchShifterNode()
